+++ ibm_mq/datadog_checks/ibm_mq/stats/channel_stats.py
@@ -4,7 +4,7 @@
 from pymqi.CMQC import MQCA_REMOTE_Q_MGR_NAME
 from pymqi.CMQCFC import MQCACH_CHANNEL_NAME, MQCACH_CONNECTION_NAME, MQGACF_CHL_STATISTICS_DATA, MQIACH_CHANNEL_TYPE
 
-from datadog_checks.ibm_mq.stats_wrapper.base_stats import BaseStats
+from datadog_checks.ibm_mq.stats.base_stats import BaseStats
 
 from ..utils import sanitize_strings
